+++ packages/create-game/eslint.config.js
@@ -1,0 +1,14 @@
+import browserConfig from "@inglorious/eslint-config/browser"
+import nodeConfig from "@inglorious/eslint-config/node"
+import { defineConfig } from "eslint/config"
+
+export default defineConfig([
+  ...browserConfig,
+  ...nodeConfig,
+  {
+    files: ["./bin/index.js"],
+    rules: {
+      "no-console": "off",
+    },
+  },
+])