+++ packages/create-game/templates/ijs/vite.config.js
@@ -1,0 +1,15 @@
+import { defineConfig } from "vite"
+import babel from "vite-plugin-babel"
+
+export default defineConfig({
+  plugins: [
+    babel({
+      include: "src/**",
+      filter: /\.(js|ijs)$/,
+
+      babelConfig: {
+        presets: ["@inglorious/inglorious-script"],
+      },
+    }),
+  ],
+})