+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @inglorious/docs
 
+## 1.7.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/eslint-config@1.0.0
+  - @inglorious/engine@2.1.3
+  - @inglorious/renderer-2d@0.8.3
+  - @inglorious/renderer-react-dom@0.6.3
+  - @inglorious/utils@3.5.3
+
 ## 1.7.3
 
 ### Patch Changes